+++ examples/knobs_example/lib/home_page.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:widgetbook/widgetbook.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage({
@@ -9,6 +8,7 @@ class HomePage extends StatefulWidget {
     this.countLabel,
     this.iconData,
     this.showToolTip = true,
+    this.duration = Duration.zero,
     this.dateTime,
   });
 
@@ -17,6 +17,7 @@ class HomePage extends StatefulWidget {
   final String? countLabel;
   final IconData? iconData;
   final bool showToolTip;
+  final Duration duration;
   final DateTime? dateTime;
 
   @override
@@ -44,9 +45,13 @@ class _HomePageState extends State<HomePage> {
               '$_counter',
               style: Theme.of(context).textTheme.headlineMedium,
             ),
+            Text(
+              '${widget.duration}',
+              style: Theme.of(context).textTheme.headlineMedium,
+            ),
             if (widget.dateTime != null)
               Text(
-                DateTimeField.asString(widget.dateTime!),
+                widget.dateTime!,
                 style: Theme.of(context).textTheme.headlineMedium,
               ),
           ],