+++ packages/widgetbook/lib/src/fields/field_type.dart
@@ -5,5 +5,6 @@ enum FieldType {
   doubleInput,
   list,
   string,
+  duration,
   dateTime,
 }