+++ packages/widgetbook_cli/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - **BREAKING**: Remove `--git-provider` option from `publish` command. ([#913](https://github.com/widgetbook/widgetbook/pull/913))
+- **BREAKING**: Remove `--base-commit` option from `publish` command. ([#936](https://github.com/widgetbook/widgetbook/pull/936))
 
 ## 3.0.0-rc.3